tools/scraper_requirements.txt
# Web Scraper Dependencies
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
html5lib>=1.1
mem0ai>=1.0.0
